+++ src/briefy/reflex/__init__.py
@@ -1,4 +1,9 @@
 """briefy.reflex."""
+from briefy import common
+from briefy import gdrive
+from briefy import reflex
+from zope.configuration.xmlconfig import XMLConfig
+
 import logging
 
 
@@ -8,3 +13,8 @@ logger.setLevel(logging.INFO)
 cs = logging.StreamHandler()
 cs.setLevel(logging.INFO)
 logger.addHandler(cs)
+
+
+XMLConfig('configure.zcml', common)()
+XMLConfig('configure.zcml', gdrive)()
+XMLConfig('configure.zcml', reflex)()